make the submit history default startdate the first date a form was submitted
